UBUNTU: backports: import HCI SMD driver
We need to carry and build the HCI SMD driver from now on as part of the backports tree to
keep the dependency on the right bluetooth layer.

The driver itself copied and modified version of the one still shipped as part of the
older bluetooth stack we have in the tree under net/bluetooth. Code was taken from
drivers/bluetooth/hci_smd.c and adjusted to fit the changed HCI driver API.

BugLink: http://bugs.launchpad.net/bugs/1489327

Signed-off-by: Simon Fels <simon.fels@canonical.com>